requirements.txt
asgiref
Django~=5.0.7
sqlparse
python-dotenv~=1.0.1
path
psycopg2-binary
djangorestframework~=3.15.2
Pillow

django-filter~=24.2